python_tutorial/pyspark_tutorial/pyspark_project/batch-etl-pyspark/requirements.txt
pyspark
python-dotenv
apache-airflow
pytest
pandas
